Added armouring stream; Moved armour-related code to stream-armour.c/.h
